+++ data-raw/ASGS_Ed3.R
@@ -17,7 +17,7 @@ library(nngeo)
 sf_use_s2(FALSE)
 
 # reduce the resolution of the borders to 1km
-tolerance_m <- 1000L
+tolerance_m <- 750L
 
 # list layers available
 st_layers(abs_geopackage)
@@ -27,44 +27,79 @@ lga <- read_sf(abs_geopackage, layer = "LGA_2021_AUST_GDA2020")
 crs_nsw <- sf::st_crs(7844) # GDA2020
 crs_working <- sf::st_crs("+proj=eqc +lat_ts=34 units=m")
 
-# Unincorporated NSW includes Lord Howe Island but also the Unincorporated Far
-# West Region. Remove LHI by culling everything east of 150.
+# NSW administrative boundaries have some quirks to account for:
+#
+#   - Unincorporated Far West Region
+#   - Lord Howe Island
+#   - Australian Capital Territory
+#   - Jervis Bay Territory
+
+# The ABS LGA region of "Unincorporated NSW" includes both Lord Howe Island
+# and the Unincorporated Far West Region. Separate these.
 nsw_hires_ui <- filter(lga, LGA_NAME_2021 == "Unincorporated NSW")
 bb <- st_bbox(nsw_hires_ui)
 bb["xmax"] <- 150
-nsw_hires_ui <- st_intersection(
+ufwr_hires <- st_intersection(
   st_transform(nsw_hires_ui, crs_working),
   st_transform(st_as_sfc(bb), crs_working)
-) %>% st_transform(crs_nsw)
+) |> st_transform(crs_nsw)
+bb <- st_bbox(nsw_hires_ui)
+bb["xmin"] <- 158
+lhi_hires <- st_intersection(
+  st_transform(nsw_hires_ui, crs_working),
+  st_transform(st_as_sfc(bb), crs_working)
+) |> st_transform(crs_nsw)
+lhi <- st_geometry(lhi_hires)
+object.size(lhi)
+usethis::use_data(lhi, overwrite = TRUE)
 
-lga_nsw <- lga %>%
-  filter(STATE_NAME_2021 == "New South Wales", LGA_NAME_2021 != "Unincorporated NSW") %>%
-  rbind(nsw_hires_ui) %>%
-  st_transform(crs_working) %>%
-  st_simplify(dTolerance = tolerance_m) %>%
+# The ABS LGA region of "Unincorp. Other Territories" includes both Norfolk
+# Island and the Jervis Bay Territory, of which we only need the latter.
+aus_hires_ui <- filter(lga, LGA_NAME_2021 == "Unincorp. Other Territories")
+bb <- st_bbox(aus_hires_ui)
+bb["xmax"] <- 153
+jbt_hires <- st_intersection(
+  st_transform(aus_hires_ui, crs_working),
+  st_transform(st_as_sfc(bb), crs_working)
+) |> st_transform(crs_nsw)
+jbt <- st_geometry(jbt_hires)
+object.size(jbt)
+usethis::use_data(jbt, overwrite = TRUE)
+
+lga_nsw <- lga |>
+  filter(STATE_NAME_2021 == "New South Wales", LGA_NAME_2021 != "Unincorporated NSW") |>
+  rbind(ufwr_hires) |>
+  st_transform(crs_working) |>
+  st_simplify(dTolerance = tolerance_m) |>
   st_transform(crs_nsw)
 object.size(lga_nsw)
 usethis::use_data(lga_nsw, overwrite = TRUE)
 
-nsw_hires <- lga %>%
-  filter(STATE_NAME_2021 == "New South Wales", LGA_NAME_2021 != "Unincorporated NSW") %>%
-  rbind(nsw_hires_ui) %>%
-  st_transform(crs_working) %>%
-  st_union() %>%
-  st_remove_holes() %>%
+act_hires <- lga |>
+  filter(STATE_NAME_2021 == "Australian Capital Territory", ! st_is_empty(geom))
+act <- st_geometry(act_hires)
+object.size(act)
+usethis::use_data(act, overwrite = TRUE)
+
+nsw_hires <- lga |>
+  filter(STATE_NAME_2021 == "New South Wales", LGA_NAME_2021 != "Unincorporated NSW") |>
+  rbind(ufwr_hires, jbt_hires) |>
+  st_transform(crs_working) |>
+  st_union() |>
+  st_remove_holes() |>
   st_make_valid()
-nsw <- nsw_hires %>%
-  st_simplify(dTolerance = tolerance_m) %>%
+nsw <- nsw_hires |>
+  st_simplify(dTolerance = tolerance_m) |>
   st_transform(crs_nsw)
 object.size(nsw_hires)
 object.size(nsw)
 usethis::use_data(nsw, overwrite = TRUE)
 
 # sal <- read_sf(abs_geopackage, layer = "SAL_2021_AUST_GDA2020")
-# sal_nsw <- sal %>%
-#   st_transform(crs_working) %>%
-#   filter(STATE_NAME_2021 == "New South Wales") %>%
-#   st_simplify(dTolerance = tolerance_m) %>%
+# sal_nsw <- sal |>
+#   st_transform(crs_working) |>
+#   filter(STATE_NAME_2021 == "New South Wales") |>
+#   st_simplify(dTolerance = tolerance_m) |>
 #   st_transform(crs_nsw)
 # object.size(sal_nsw)
 
@@ -78,10 +113,10 @@ poa_in_nsw <- st_intersects(
   st_buffer(nsw_hires, -100L)) |>
   sapply(function(x) length(x) > 0)
 
-poa_nsw <- poa[poa_in_nsw, ] %>%
-  st_transform(crs_working) %>%
-  st_simplify(dTolerance = tolerance_m, preserveTopology = TRUE) %>%
-  st_transform(crs_nsw) %>%
+poa_nsw <- poa[poa_in_nsw, ] |>
+  st_transform(crs_working) |>
+  st_simplify(dTolerance = tolerance_m, preserveTopology = TRUE) |>
+  st_transform(crs_nsw) |>
   st_make_valid()
 stopifnot(all(!st_is_empty(poa_nsw)))
 stopifnot(all(st_geometry_type(poa_nsw) != "GEOMETRYCOLLECTION"))