+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * Add Primary Health Network (PHN) map data as `phn` dataset.
 * Fixed the `poa_nsw` dataset to avoid strange geometries near the state border.
+* Added geometries for Lord Howe Island, Australian Capital Territory, and
+  Jervis Bay Territory, and a utility function `outline()` for using them.
+* Slightly increase the resolution to 750 m instead of 1 km.
 
 # nswgeo 0.3.3